Optimize DictionaryWriteHandler.StringableKeys w/ static knowledge
When writing a dictionary where the key type is defined dynamically, and the key tag is always the same for any non-null value, the KnownTag can be checked without a full scan of the dictionary keys.
I didn't really find an automatic way to test this optimization, especially since if it doesn't short-circuit, the full scan occurs and it gets the right answer anyway.