Performance boost: Utf8JsonParser no longer performs test read
The Utf8JsonStreamReader kept its own state and read the token data twice every time so it could return a whole token.  By pushing some complexity back onto the Utf8JsonParser it only has to read once in the happy path when data is available.
This also returned a byte sequence that is redundant because it's already a member of the rdr.