Test some UNICODE characters
According to my TimeExecution.Program.DoRune(), there aren't currently any 4-UTF8-byte unicode characters that contain the UTF-8 '\\' character used to initiate a JSON string escape sequence.
A note about the chosen characters:
Historically, the only encoding supported by the Newtonsoft.Json-related TransitFactory was Encoding.Default.
On my machine, the default codeset (used by Encoding.Default) is Windows-1252. In some cases, these characters are munged, but not in others.  As such, some of the "expected" test strings needed to be Transcoded to agree with the limitation, however it may be that different codepages will need different transcodings encodings and thus my list of "should be transcoded expected strings" may be incomplete.
I expect that the long-term solution to this problem is to parameterize the encoding when creating the reader and writer, but that will wait for another commit.